feat: add game restart button
This commit adds a "restart game" button, to make it easier to load
complete save file histories after modifying the save directory.
The commit adds a new user-configurable setting to define a shell
command to restart the game. The setting is pre-configured with
shell commands for the major distribution platforms (Steam and Game
Pass), but it also allows the user to enter an arbitrary command.