feat: GUI revamp and support folders
This commit significantly changes the GUI. Most notably, it adds
support for folder management, allowing saves to be organized into
subfolders within the same profile. To accomplish this, the commit
includes substantial changes to the save file manager and the
replacement of the QListView with a QTreeView backed by a custom
QAbstractItemModel.

The commit also includes various quality-of-life improvements, both
automatic and configurable in the settings window. There is a new
option for expanding all folders on startup, and another option to
control the double-click action for savestates (notably, it is now
possible to load the save and restart the game with a double-click
on a savestate). The window geometry is also now saved to the
config file automatically, restoring the window position and size
on startup. The commit also adds a "replace savestate" button (in
addition to the context menu), and various buttons and features
related to folder management.

This commit does not include unit tests for the new functionality,
and removes obsolete ones. These can be re-added in the future if
the need for production-ready code arises.